# Evaluate lookup tables in compile time