--- a/jsoniter-scala-core/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonWriter.scala
+++ b/jsoniter-scala-core/src/main/scala/com/github/plokhotnyuk/jsoniter_scala/core/JsonWriter.scala
@@ -1870,11 +1870,14 @@ final class JsonWriter private[jsoniter_scala](
         if (e2 >= 0) {
           val ss = f64Pow5InvSplit
           val q = Math.max(0, (e2 * 1292913986L >> 32).toInt - 1) // == Math.max(0, (e2 * Math.log10(2)).toInt - 1)
-          val i = -e2 + q + Math.max(0, (q * 9972605231L >> 32).toInt) + 122 // == -e2 + q + Math.max(0, (q * Math.log(5) / Math.log(2)).toInt) + 122
+          val idx = q << 1
+          val s0 = ss(idx)
+          val s1 = ss(idx + 1)
+          val j = -e2 + q + Math.max(0, (q * 9972605231L >> 32).toInt) + 122 // == -e2 + q + Math.max(0, (q * Math.log(5) / Math.log(2)).toInt) + 122
           exp = q
-          dv = fullMulPow5DivPow2(mv, q, i, ss)
-          dp = fullMulPow5DivPow2(mp, q, i, ss)
-          dm = fullMulPow5DivPow2(mm, q, i, ss)
+          dv = fullMulPow5DivPow2(mv, j, s0, s1)
+          dp = fullMulPow5DivPow2(mp, j, s0, s1)
+          dm = fullMulPow5DivPow2(mm, j, s0, s1)
           if (q <= 21) {
             val mv5 = mv / 5
             if ((mv5 << 2) + mv5 == mv) dvIsTrailingZeros = multiplePowOf5(mv5, q - 1)
@@ -1885,11 +1888,14 @@ final class JsonWriter private[jsoniter_scala](
           val ss = f64Pow5Split
           val q = Math.max(0, (-e2 * 3002053309L >> 32).toInt - 1) // == Math.max(0, (-e2 * Math.log10(5)).toInt - 1)
           val i = -e2 - q
+          val idx = i << 1
+          val s0 = ss(idx)
+          val s1 = ss(idx + 1)
           val j = q - Math.max(0, (i * 9972605231L >> 32).toInt) + 120 // == q - Math.max(0, (i * Math.log(5) / Math.log(2)).toInt) + 120
           exp = -i
-          dv = fullMulPow5DivPow2(mv, i, j, ss)
-          dp = fullMulPow5DivPow2(mp, i, j, ss)
-          dm = fullMulPow5DivPow2(mm, i, j, ss)
+          dv = fullMulPow5DivPow2(mv, j, s0, s1)
+          dp = fullMulPow5DivPow2(mp, j, s0, s1)
+          dm = fullMulPow5DivPow2(mm, j, s0, s1)
           if (q <= 1) {
             dvIsTrailingZeros = true
             if (even) dmIsTrailingZeros = mmShift == 1
@@ -2020,16 +2026,15 @@ final class JsonWriter private[jsoniter_scala](
     (q1 << 2) + q1 == q0 && multiplePowOf5(q1, q - 1)
   }
 
-  private[this] def fullMulPow5DivPow2(m: Long, i: Int, j: Int, ss: Array[Int]): Long = {
+  private[this] def fullMulPow5DivPow2(m: Long, j: Int, s0: Long, s1: Long): Long = {
     val ml = m & 0x7FFFFFFF
     val mh = m >>> 31
-    val idx = i << 2
-    val s3 = ss(idx + 3)
-    val s2 = ss(idx + 2)
-    val s1 = ss(idx + 1)
-    val s0 = ss(idx)
-    ((((((((ml * s3 >>> 31) + ml * s2 + mh * s3) >>> 31) + ml * s1 +
-      mh * s2) >>> 31) + ml * s0 + mh * s1) >>> 21) + (mh * s0 << 10)) >>> (j - 114)
+    val s0l = s0 & 0x7FFFFFFF
+    val s0h = s0 >>> 31
+    val s1l = s1 & 0x7FFFFFFF
+    val s1h = s1 >>> 31
+    ((((((((ml * s0l >>> 31) + ml * s0h + mh * s0l) >>> 31) + ml * s1l +
+      mh * s0h) >>> 31) + ml * s1h + mh * s1l) >>> 21) + (mh * s1h << 10)) >>> (j - 114)
   }
 
   private[this] def offset(q0: Long): Int = {
@@ -2192,34 +2197,28 @@ object JsonWriter {
     }
     fs
   }
-  private final val f64Pow5InvSplit: Array[Int] = eval {
-    val fs = new Array[Int](1164)
+  private final val f64Pow5InvSplit: Array[Long] = eval {
+    val fs = new Array[Long](582)
     var pow5 = BigInt(1)
     var i = 0
-    while (i < 291) {
+    while (i < 582) {
       val inv = ((BigInt(1) << (pow5.bitLength + 121)) / pow5) + 1
-      var j = 0
-      while (j < 4) {
-        fs(i * 4 + j) = (inv >> ((3 - j) * 31)).intValue & 0x7FFFFFFF
-        j += 1
-      }
+      fs(i) = inv.longValue & 0x3FFFFFFFFFFFFFFFL
+      fs(i + 1) = (inv >> 62).longValue & 0x3FFFFFFFFFFFFFFFL
       pow5 *= 5
-      i += 1
+      i += 2
     }
     fs
   }
-  private final val f64Pow5Split: Array[Int] = eval {
-    val fs = new Array[Int](1304)
+  private final val f64Pow5Split: Array[Long] = eval {
+    val fs = new Array[Long](652)
     var pow5 = BigInt(1)
     var i = 0
-    while (i < 326) {
-      var j = 0
-      while (j < 4) {
-        fs(i * 4 + j) = (pow5 >> (pow5.bitLength - 121 + (3 - j) * 31)).intValue & 0x7FFFFFFF
-        j += 1
-      }
+    while (i < 652) {
+      fs(i) = (pow5 >> (pow5.bitLength - 121)).longValue & 0x3FFFFFFFFFFFFFFFL
+      fs(i + 1) = (pow5 >> (pow5.bitLength - 59)).longValue & 0x3FFFFFFFFFFFFFFFL
       pow5 *= 5
-      i += 1
+      i += 2
     }
     fs
   }