It allows to speed up startup of parser and serializer. Parsing of timezone still affected by delay due initialization of cached values for them.

Bellow is timing of starting jsoniter-scala-example using JDK 8 and native image compiled by GraalVM CE 19 accordingly.

Before:
```
s$ time java -jar target/scala-2.12/jsoniter-scala-examples-assembly-0.1.0-SNAPSHOT.jar
User(John,List(Device(1,HTC One X)))
{"name":"John","devices":[{"id":2,"model":"iPhone X"}]}

real	0m0,376s
user	0m0,814s
sys	0m0,044s

$ time ./jsoniter-scala-examples-assembly-0.1.0-SNAPSHOT
User(John,List(Device(1,HTC One X)))
{"name":"John","devices":[{"id":2,"model":"iPhone X"}]}

real	0m0,013s
user	0m0,014s
sys	0m0,000s
```

After:
```
$ time java -jar target/scala-2.12/jsoniter-scala-examples-assembly-0.1.0-SNAPSHOT.jar
User(John,List(Device(1,HTC One X)))
{"name":"John","devices":[{"id":2,"model":"iPhone X"}]}

real	0m0,323s
user	0m0,391s
sys	0m0,027s

$ time ./jsoniter-scala-examples-assembly-0.1.0-SNAPSHOT
User(John,List(Device(1,HTC One X)))
{"name":"John","devices":[{"id":2,"model":"iPhone X"}]}

real	0m0,001s
user	0m0,001s
sys	0m0,000s
```